Fix IO Node Selection Not Maintained when Name is Updated (#846)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->

Fixes an issue where save-on-blur functionality was causing updates to IO Node Names to become out-of-sync between the canvas, spec and context panel; resulting in selection issues and save issues.

The selection state will now be migrated across to the new node using a slightly-hacky workaround. A better solution can hopefully be found in future.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->
Progresses https://github.com/Shopify/oasis-frontend/issues/239

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Bug fix
- [ ] New feature
- [ ] Improvement
- [ ] Cleanup/Refactor
- [ ] Breaking change
- [ ] Documentation update

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->

The solution here, with a zero-second timeout is less than ideal, but still better than having it not working at all. This is something that we should investigate further in the future to try find a more sustainable longer-term solution.