InputEditor Default Value <> Value (#986)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
Users are getting confused about the InputValueEditor `default value` field. Since subgraphs are not formally supported nearly all users are using this as a `value` field and not a `default value field`. This currently has no impact on the operation of pipelines since we mirror the default value with the value. However, the terminology does cause confusion for users.

This PR aims to try provide a stop-gap while subgraphs are implemented by renaming the field to `value` and indicating that it will also be used as the default.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->
Closes https://github.com/Shopify/oasis-frontend/issues/277

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Improvement

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

![image.png](https://app.graphite.dev/user-attachments/assets/07b2764e-2dcd-41b2-b135-3fb017328209.png)

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->