Resolve Issues with useLoadComponentSpecFromPath making permanent changes when checking a url (#988)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
Resolves an issue introduced in #970 that resulted in a Run's editor version of the component spec being loaded. This resulted in tasks on the runs having missing (or outdated) execution ids and thus missing (or outdated) artifacts and logs.

This PR resolves that by separating out the `check if spec exists` use case from the original `load the spec` use case. In particular, the issue was caused by `setComponentSpec` being trigged when a url was checked.

Adds tests to the PipelineRun Route to check that tasks have execution ids attached to them.

**Original Issue**
In PR #970 `useLoadComponentSpecFromPath` is being used to test a url and see if the editor route from a pipeline can be loaded. I added some escape hatches to the hook for this scenario to avoid any permanent state changes but missed two cases of `setComponentSpec`. This resulted in the run's spec  being overwritten with the editor's version of the spec (which does not have execution ids).

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

this should resolve the underlying issues for https://github.com/Shopify/oasis-frontend/issues/280 and https://github.com/Shopify/oasis-frontend/issues/281

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Bug fix

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
1. run a pipeline
2. task details, artifacts and logs should be viewable (previously not)
3. inspect pipeline button should not be viewable if you don't have the pipeline saved in your browser (intended behaviour introduced by #970 )

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->