Remove PipelineRunsProvider from Run Route (#910)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->

Adds native PipelineRun "rerun" submission logic, removing the need to call in `PipelineRunsProvider`.

`PipelineRunsProvider` is a provider intended to provide the `editor` route with details and management tools for pipeline runs created on that pipeline. It was never intended to be used within the `run` route. Calling it in results in a large number of pointless backend queries to attain a list of runs that is completely unused on this route. Hence, it makes sense to move out the part that we specifically need, `submit`, and ditch the provider.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

Progresses: https://github.com/Shopify/oasis-frontend/issues/171

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Cleanup/Refactor

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
Can re-run pipelines as normal.
Should see less backend queries when loading a pipeline run.

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->