Load Pipeline Run by `run_id` or `root_execution_id` (#1051)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
Supersedes #859, #1045 and remedies the issues that caused the rollback in #1052

Allows users to navigate to a pipeline run using the `run_id` instead of the `root_execution_id` if they choose to. This logic is handled through a new hook that will attempt to fetch data from an input `id` first assuming it is a `root_execution_id` and, if not, a `run_id`.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->
Partially addresses https://github.com/Cloud-Pipelines/pipeline-studio-app/issues/128

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Bug fix
- [x] Improvement

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
Verify you can navigate to and from pipeline runs via the run id, and also still use the clone & inspect & rerun buttons as expected.

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->