Add Cache Toggle for Tasks (#1223)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
Reworks the Annotation tab of the task details into "Configuration" and adds a section to toggle Task Caching on or off. Includes a little more refactoring/moving/renaming of code, just so things make more sense.

This toggle goes through the standard Task Callback process to set `task_spec.execution_options.caching_strategy.max_cache_staleness = "P0D"`, which is currently the only accepted value on the backend. This can be expanded in future when the backend supports it.

Note: the backend does not support caching strategies for Subgraph Nodes, so the interface has been hidden in that scenario. Additionally, due to overlap with #1185 the ability to toggle cache on nested components in a subgraph will come as a follow-up once #1185 is merged.

Tasks with caching disabled will run fresh when the pipeline is executed!

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

Closes https://github.com/Shopify/oasis-frontend/issues/223

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] New feature

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

![image.png](https://app.graphite.dev/user-attachments/assets/edc02bf5-55f4-458d-ad25-1e388e3b243e.png)

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
Note: this requires the backend is updated to at least commit https://github.com/Cloud-Pipelines/backend/commit/5b4bdd5c37e3c386b569d2fdc7e9cdd2ce3bc23d

1. Create a basic pipeline
2. Run it
3. When finished, run it again and confirm it is using the cache
4. Go into the pipeline, select a task, navigate to the new "Configuration" tab and disabled the cache via the toggle.
5. Run the pipeline again and confirm that the task did not use the cache this time
6. Go back again, reactive cache, run the pipeline and confirm the task now uses cache again

Note: cache strategies are disabled for subgraph nodes and are not (yet) possible within subgraphs

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->