Refactor Annotation Editor Logic (#1201)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->

Reworks the BUSINESS LOGIC ONLY of the annotation editor so that it is more in-line with our current practices in regard to input editing and autosave. In other words, it removes all of the external `onChange` handlers that were no longer relevant following the move to save-on-blur (but never cleaned up). The current value of the input field is now tracked internally and committed when blurred. This means it now also works with similar structure to the argument editor; thus enabling us to add save-on-unmount functionality in an upstack PR.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->

Makes some progress toward https://github.com/Shopify/oasis-frontend/issues/235

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Cleanup/Refactor

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->

App should behave unchanged from previous.

Editing, viewing, creating, deleting annotations and compute resources should all work without change or issue.

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->