Fix ReactKey issues with multiple new annotation rows (#1240)
## Description

<!-- Please provide a brief description of the changes made in this pull request. Include any relevant context or reasoning for the changes. -->
Fixes a bug where React Keys for new annotation rows were unstable due to being based on the row `index` in the array. This would cause issues when, for example, a row was delete or submitted, as all other rows would shift upwards and inherit the info of the row above.

The solution is to assign each new row a unique id specifically for use by React.


I also did a brief bit of tidying up while I was there, by introducing a proper `NewRow` type.

## Related Issue and Pull requests

<!-- Link to any related issues using the format #<issue-number> -->
Closes https://github.com/Shopify/oasis-frontend/issues/336

## Type of Change

<!-- Please delete options that are not relevant -->

- [x] Bug fix

## Checklist

<!-- Please ensure the following are completed before submitting the PR -->

- [ ] I have tested this does not break current pipelines / runs functionality
- [ ] I have tested the changes on staging

## Screenshots (if applicable)

<!-- Include any screenshots that might help explain the changes or provide visual context -->

## Test Instructions

<!-- Detail steps and prerequisites for testing the changes in this PR -->
- Confirm you can now edit/delete/navigate multiple new annotation rows without results being mixed up or erroneously deleted.

## Additional Comments

<!-- Add any additional context or information that reviewers might need to know regarding this PR -->